+++ lib/i18n.d.ts
@@ -1,2 +1,0 @@
-export const T: (key: import("i18next").TFunctionKeys, options: import("i18next").TOptions<import("i18next").StringMap>, lng: string) => import("i18next").TFunctionDetailedResult<object>;
-export const TT: (locale: string) => (key: import("i18next").TFunctionKeys, options: import("i18next").TOptions<import("i18next").StringMap>) => import("i18next").TFunctionDetailedResult<object>;